rename: HotPlace로 시작하는 파일 Restaurant로 변경
